requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-dotenv==1.0.1
psycopg[binary]==3.2.3
python-multipart==0.0.9
